requirements.txt
nuitka==2.6.9
pywebview==5.4